feat: use normal dialogs when targetting body (intro steps)
