+++ frontend/src/component/demo/DemoSteps/DemoSteps.tsx
@@ -3,64 +3,11 @@ import Joyride, {
     CallBackProps,
     TooltipRenderProps,
 } from 'react-joyride';
-import {
-    Button,
-    IconButton,
-    Typography,
-    alpha,
-    styled,
-    useTheme,
-} from '@mui/material';
-import CloseIcon from '@mui/icons-material/Close';
+import { useTheme } from '@mui/material';
 import { ITutorialTopic, ITutorialTopicStep } from '../demo-topics';
 import { useEffect, useState } from 'react';
-import { ConditionallyRender } from 'component/common/ConditionallyRender/ConditionallyRender';
 import { useLocation, useNavigate } from 'react-router-dom';
-
-const StyledTooltip = styled('div')(({ theme }) => ({
-    '@keyframes pulse': {
-        '0%': {
-            boxShadow: `0 0 0 0 ${alpha(theme.palette.primary.main, 0.7)}`,
-        },
-        '70%': {
-            boxShadow: `0 0 0 10px ${alpha(theme.palette.primary.main, 0)}`,
-        },
-        '100%': {
-            boxShadow: `0 0 0 0 ${alpha(theme.palette.primary.main, 0)}`,
-        },
-    },
-    position: 'relative',
-    backgroundColor: theme.palette.background.paper,
-    color: theme.palette.text.primary,
-    borderRadius: theme.shape.borderRadiusMedium,
-    width: '100%',
-    maxWidth: theme.spacing(45),
-    padding: theme.spacing(3),
-}));
-
-const StyledCloseButton = styled(IconButton)(({ theme }) => ({
-    position: 'absolute',
-    right: theme.spacing(1),
-    top: theme.spacing(1),
-    color: theme.palette.neutral.main,
-}));
-
-const StyledTooltipTitle = styled('div')(({ theme }) => ({
-    display: 'flex',
-    alignItems: 'center',
-    gap: theme.spacing(1),
-    marginBottom: theme.spacing(1),
-    flexWrap: 'wrap',
-}));
-
-const StyledTooltipActions = styled('div')(({ theme }) => ({
-    display: 'flex',
-    justifyContent: 'space-between',
-    marginTop: theme.spacing(3),
-    '&&& button': {
-        fontSize: theme.fontSizes.smallBody,
-    },
-}));
+import { DemoStepTooltip } from './DemoStepTooltip/DemoStepTooltip';
 
 interface IDemoStepsProps {
     setExpanded: React.Dispatch<React.SetStateAction<boolean>>;
@@ -262,62 +209,20 @@ export const DemoSteps = ({
                     mixBlendMode: 'unset',
                 },
             }}
-            tooltipComponent={({
-                step,
-                tooltipProps,
-            }: TooltipRenderProps & {
-                step: ITutorialTopicStep;
-            }) => (
-                <StyledTooltip {...tooltipProps}>
-                    <StyledCloseButton aria-label="close" onClick={close}>
-                        <CloseIcon />
-                    </StyledCloseButton>
-                    <StyledTooltipTitle>
-                        <ConditionallyRender
-                            condition={Boolean(step.title)}
-                            show={step.title}
-                            elseShow={
-                                <Typography fontWeight="bold">
-                                    {topics[topic].title}
-                                </Typography>
-                            }
-                        />
-                    </StyledTooltipTitle>
-                    {step.content}
-                    <StyledTooltipActions>
-                        <div>
-                            <ConditionallyRender
-                                condition={topic > 0 || steps[topic] > 0}
-                                show={
-                                    <Button
-                                        variant="outlined"
-                                        onClick={() => onBack(step)}
-                                    >
-                                        Back
-                                    </Button>
-                                }
-                            />
-                        </div>
-                        <div>
-                            <ConditionallyRender
-                                condition={Boolean(step.nextButton)}
-                                show={
-                                    <Button
-                                        onClick={() => next(steps[topic])}
-                                        variant="contained"
-                                        sx={{ alignSelf: 'flex-end' }}
-                                    >
-                                        {topic === topics.length - 1 &&
-                                        steps[topic] ===
-                                            topics[topic].steps.length - 1
-                                            ? 'Finish'
-                                            : 'Next'}
-                                    </Button>
-                                }
-                            />
-                        </div>
-                    </StyledTooltipActions>
-                </StyledTooltip>
+            tooltipComponent={(
+                props: TooltipRenderProps & {
+                    step: ITutorialTopicStep;
+                }
+            ) => (
+                <DemoStepTooltip
+                    {...props}
+                    topic={topic}
+                    topics={topics}
+                    steps={steps}
+                    onClose={close}
+                    onBack={onBack}
+                    onNext={next}
+                />
             )}
         />
     );
